% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Wrap paste() with collapse argument}
\usage{
collapse(vec, sep = "")
}
\arguments{
\item{vec}{string vector}

\item{sep}{separator}
}
\value{
string with length 1
}
\description{
Wrap paste() with collapse argument
}
\examples{
collapse(c("a", "b", "c"), "-")
}
