% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brackets.R
\name{[<-}
\alias{[<-}
\title{Mask \verb{[<-} to allow character assignment
See [base::\verb{[<-}()]}
\usage{
`[`(x, i, ...) <- value
}
\arguments{
\item{x}{object}

\item{i}{index}

\item{...}{ignored / passed on}

\item{value}{value to assign}
}
\description{
Mask \verb{[<-} to allow character assignment
See [base::\verb{[<-}()]
}
